Adds in armor-based objectives (SoH, FoTL, etc) to pursuits display
This will allow armor pieces that have objectives tied into them (Such as the armor pieces from the "Solstice of Heroes" event) to be displayed as a pursuit.

This change also removes pursuits from being displayed if they do not have any objectives tied to them. These are generally 'marker' items for the player, and do not need to be displayed here.